services/credit/pyproject.toml
[project]
name = "prequal-credit"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.120.1,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "kafka-python (>=2.2.15,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.14.2,<0.15.0)",
    "pre-commit (>=4.3.0,<5.0.0)"
]
